requirements.txt
discord.py
asyncio~=3.4.3
discord-py-slash-command
datetime~=4.4
mysql~=0.0.3
mysql-connector-python==8.0.29
pandas~=1.4.2
discord~=1.7.3
PyYAML~=6.0